make a service out of chatAttachmentResolve.js
